Finished Fixing UI Resizing
Still needs to be tested on various other resolutions and screen sizes.
