+++ app/src/main/java/kidsense/kadho/com/kidsense_offline_demo/featureControl.java
@@ -19,16 +19,6 @@ import java.util.Arrays;
  * A simple {@link Fragment} subclass.
  */
 public class featureControl extends Fragment {
-    private ArrayList<String> authoizedWebsites = new ArrayList<String>(Arrays.asList("WWW.KIDSESNE.AI", "WWW.KADHO.COM"));
-
-    private int deleteWebsite = -1;
-    private ArrayAdapter<String> listAdapter;
-
-    EditText text;
-    Button addButton;
-    Button deleteButton;
-    ListView listView;
-
     public featureControl() {
         // Required empty public constructor
     }
@@ -41,73 +31,6 @@ public class featureControl extends Fragment {
 
         View rootView = inflater.inflate(R.layout.tab_feature_control, container, false);
 
-        text = (EditText) rootView.findViewById(R.id.enterWebsite);
-        addButton = (Button) rootView.findViewById(R.id.addWebsiteButton);
-        deleteButton = (Button) rootView.findViewById(R.id.deleteWebsiteButton);
-        listView = (ListView) rootView.findViewById(R.id.allWebsites);
-
-        listView = (ListView)rootView.findViewById(R.id.allWebsites);
-        listAdapter = new ArrayAdapter<String>(featureControl.this.getActivity(), android.R.layout.simple_list_item_multiple_choice, authoizedWebsites);
-        listView.setChoiceMode(ListView.CHOICE_MODE_SINGLE);
-        listView.setAdapter(listAdapter);
-
-
-        unMarkItems();
-        onItemClick();
-        onBtnClick();
-        onDeleteBtnClicked();
-
         return rootView;
     }
-
-    public void unMarkItems(){
-        for(int i = 0; i < authoizedWebsites.size(); i++)
-            listView.setItemChecked(i, false);
-    }
-
-    public void onItemClick(){
-        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                if(deleteWebsite == position) {
-                    listView.setItemChecked(deleteWebsite, false);
-                    deleteWebsite = -1;
-                }
-                else
-                    deleteWebsite = position;
-            }
-        });
-    }
-
-    public void onBtnClick(){
-        addButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                String website = text.getText().toString().toUpperCase();
-                if(!website.isEmpty() && !authoizedWebsites.contains(website))
-                    authoizedWebsites.add(website);
-
-                text.setText("");
-                unMarkItems();
-                listAdapter.notifyDataSetChanged();
-
-            }
-        });
-    }
-
-    public void onDeleteBtnClicked() {
-        deleteButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if(deleteWebsite != -1) {
-                    authoizedWebsites.remove(deleteWebsite);
-                }
-
-                deleteWebsite = -1;
-                unMarkItems();
-                listAdapter.notifyDataSetChanged();
-            }
-        });
-    }
-
 }
